+++ libs/vscode/debug/src/lib/idl-debug-configuration.ts
@@ -3,6 +3,8 @@ import { CleanPath } from '@idl/shared';
 import { IDL_LOGGER } from '@idl/vscode/client';
 import { IDL_EXTENSION_CONFIG } from '@idl/vscode/config';
 import { IDLExtensionConfig } from '@idl/vscode/extension-config';
+import { VariablesReferenceSubstitution } from '@idl/vscode/shared';
+import copy from 'fast-copy';
 import * as path from 'path';
 import {
   CancellationToken,
@@ -43,7 +45,7 @@ export class IDLDebugConfigurationProvider
     const useConfig = {
       ...DEFAULT_IDL_DEBUG_CONFIGURATION,
       ...config,
-      ...{ config: IDL_EXTENSION_CONFIG as IDLExtensionConfig },
+      ...{ config: copy(IDL_EXTENSION_CONFIG) as IDLExtensionConfig },
     };
 
     // set environment
@@ -52,6 +54,17 @@ export class IDLDebugConfigurationProvider
       DEFAULT_IDL_DEBUG_CONFIGURATION.env, // from `process.env`
       IDL_EXTENSION_CONFIG.IDL.environment
     );
+
+    // do substitution for all environment variables
+    const vars = Object.keys(IDL_EXTENSION_CONFIG.IDL.environment);
+    for (let i = 0; i < vars.length; i++) {
+      useConfig.env[vars[i]] = VariablesReferenceSubstitution(
+        useConfig.env[vars[i]]
+      );
+
+      // update copied config so VSCode doesnt try to map it
+      useConfig.config.IDL.environment[vars[i]] = useConfig.env[vars[i]];
+    }
 
     // string to use to join workspace folders
     const joinchar = `${path.delimiter}+`;